pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "neochef"
version = "0.1.0"
description = "A local recipe graph: discover meals from what you already have."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.2",
    "python-dotenv>=1.0.1",
    "neo4j>=5.14.0",
    "pandas>=2.3.2",
    "rolesnap>=0.8.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
notebook = [
    "notebook>=7.0.0",
    "ipykernel>=6.0.0",
]
